Dodanie wyświetlania dostępności sprzetów do ewidencji. Dokończenie obsługi błędów.

+++ source/repos/WypozyczalniaSamochodow/WypozyczalniaSamochodow/OrdersTable.cs
@@ -20,7 +20,7 @@ namespace WypozyczalniaSamochodow
         public Nullable<System.DateTime> rentalDate { get; set; }
         public Nullable<System.DateTime> returnTerm { get; set; }
         public Nullable<System.DateTime> returnDate { get; set; }
-        public Nullable<float> discount { get; set; }
+        public Nullable<int> discount { get; set; }
         public Nullable<decimal> priceOfOrder { get; set; }
         public Nullable<int> status { get; set; }
